requirements.txt
torch>=2.0.0
transformers==4.46.3
Pillow>=9.0.0
timm
einops
